requirements.txt
discord.py
asyncio
datetime
collections